Refactor API code to separate models into different code files
