+++ Web Service Endpoint/drugscom.py
@@ -1,0 +1,22 @@
+from selenium import webdriver
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.common.exceptions import WebDriverException, TimeoutException
+
+from bs4 import BeautifulSoup
+import io
+import pandas as pd
+from selenium.webdriver.support import expected_conditions as EC
+import time
+import sys
+import re
+import numpy as np
+import os
+import glob
+import shutil
+import json
+import html
+from string import punctuation
+from collections import deque
+from dotenv import load_dotenv
