+++ Web Service Endpoint/rekog.py
@@ -1,0 +1,6 @@
+import boto3
+
+
+def post_rekog(post_params):
+    return post_params
+
